marketresearch/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
crewai[google-genai]
langchain
langchain-core
langchain-community
langchain-google-genai
requests
google-generativeai
pypdf
python-docx
openpyxl
beautifulsoup4
markdown
weasyprint
chromadb